Update lock file

+++ packages/workbox-webpack-plugin/src/generate-sw.ts
@@ -223,7 +223,7 @@ class GenerateSW {
     for (const file of files) {
       compilation.emitAsset(
         file.name,
-        new RawSource(Buffer.from(file.contents)),
+        new RawSource(Buffer.from(file.contents as Uint8Array)),
         {
           // See https://github.com/webpack-contrib/compression-webpack-plugin/issues/218#issuecomment-726196160
           minimized: config.mode === 'production',
